Divide components in front
