+++ front/src/App.vue
@@ -2,7 +2,7 @@
   <div id="app">
     <md-toolbar class="md-primary" md-elevation="1">
       <h3 class="md-title" style="flex: 1">Rehearsal</h3>
-      <md-button @click.native='setRoute("/")'>Home</md-button>
+      <md-button to="/">Home</md-button>
     </md-toolbar>
     <router-view/>
   </div>
@@ -11,15 +11,7 @@
 <script>
   export default {
     name: 'App',
-    data: () => ({
-      menuVisible: false,
-      received_messages: [],
-      connected: false
-    }),
     methods: {
-      setRoute(id) {
-        this.$router.push(id);
-      }
     }
   }
 </script>
